# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfoModel"
                }
              }
            }
          }
        },
        "tags": [
          "app"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/timestamps": {
      "get": {
        "operationId": "TimestampsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimestampModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required."
          }
        },
        "tags": [
          "timestamps"
        ]
      },
      "post": {
        "operationId": "TimestampsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimestampModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampModel"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required."
          }
        },
        "tags": [
          "timestamps"
        ]
      }
    },
    "/timestamps/{id}": {
      "get": {
        "operationId": "TimestampsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampModel"
                }
              }
            }
          },
          "400": {
            "description": "Timestamp id must be a UUID."
          },
          "401": {
            "description": "Authentication required."
          },
          "404": {
            "description": "Timestamp not found."
          }
        },
        "tags": [
          "timestamps"
        ]
      },
      "patch": {
        "operationId": "TimestampsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTimestampModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampModel"
                }
              }
            }
          },
          "400": {
            "description": "Timestamp id must be a UUID."
          },
          "401": {
            "description": "Authentication required."
          },
          "404": {
            "description": "Timestamp not found."
          }
        },
        "tags": [
          "timestamps"
        ]
      },
      "delete": {
        "operationId": "TimestampsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Timestamp deleted."
          },
          "400": {
            "description": "Timestamp id must be a UUID."
          },
          "401": {
            "description": "Authentication required."
          },
          "404": {
            "description": "Timestamp not found."
          }
        },
        "tags": [
          "timestamps"
        ]
      }
    }
  },
  "info": {
    "title": "Template API",
    "description": "Backend API for the full-stack template.",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "AppInfoModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "@template/api"
          },
          "version": {
            "type": "string",
            "example": "0.0.1"
          },
          "status": {
            "type": "string",
            "example": "ok"
          },
          "appEnv": {
            "type": "string",
            "example": "local"
          },
          "nodeEnv": {
            "type": "string",
            "example": "development"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "version",
          "status",
          "appEnv",
          "nodeEnv",
          "timestamp"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-06-16T19:30:00.000Z"
          },
          "uptimeSeconds": {
            "type": "number",
            "example": 12.345
          },
          "appEnv": {
            "type": "string",
            "example": "local"
          },
          "nodeEnv": {
            "type": "string",
            "example": "development"
          }
        },
        "required": [
          "status",
          "timestamp",
          "uptimeSeconds",
          "appEnv",
          "nodeEnv"
        ]
      },
      "TimestampModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "example": "Wrote first timestamp note"
          },
          "dateOccurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "note",
          "dateOccurredAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTimestampModel": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "example": "Wrote first timestamp note"
          },
          "dateOccurredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "note",
          "dateOccurredAt"
        ]
      },
      "UpdateTimestampModel": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "example": "Updated timestamp note"
          },
          "dateOccurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}