# apps/api/src/models/create-timestamp.model.ts
import { ApiProperty } from '@nestjs/swagger';

export class CreateTimestampModel {
  @ApiProperty({ example: 'Wrote first timestamp note' })
  note!: string;

  @ApiProperty({ format: 'date-time' })
  dateOccurredAt!: string;
}