# packages/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: string;
};

export type AppInfoModel = {
  name: string;
  version: string;
  status: string;
  appEnv: string;
  nodeEnv: string;
  timestamp: string;
};

export type HealthResponseDto = {
  status: string;
  timestamp: string;
  uptimeSeconds: number;
  appEnv: string;
  nodeEnv: string;
};

export type TimestampModel = {
  id: string;
  note: string;
  dateOccurredAt: string;
  createdAt: string;
  updatedAt: string;
};

export type CreateTimestampModel = {
  note: string;
  dateOccurredAt: string;
};

export type UpdateTimestampModel = {
  note?: string;
  dateOccurredAt?: string;
};

export type AppControllerGetInfoData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/";
};

export type AppControllerGetInfoResponses = {
  200: AppInfoModel;
};

export type AppControllerGetInfoResponse =
  AppControllerGetInfoResponses[keyof AppControllerGetInfoResponses];

export type HealthControllerGetHealthData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/health";
};

export type HealthControllerGetHealthResponses = {
  200: HealthResponseDto;
};

export type HealthControllerGetHealthResponse =
  HealthControllerGetHealthResponses[keyof HealthControllerGetHealthResponses];

export type TimestampsControllerFindAllData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/timestamps";
};

export type TimestampsControllerFindAllErrors = {
  /**
   * Authentication required.
   */
  401: unknown;
};

export type TimestampsControllerFindAllResponses = {
  200: Array<TimestampModel>;
};

export type TimestampsControllerFindAllResponse =
  TimestampsControllerFindAllResponses[keyof TimestampsControllerFindAllResponses];

export type TimestampsControllerCreateData = {
  body: CreateTimestampModel;
  path?: never;
  query?: never;
  url: "/timestamps";
};

export type TimestampsControllerCreateErrors = {
  /**
   * Authentication required.
   */
  401: unknown;
};

export type TimestampsControllerCreateResponses = {
  201: TimestampModel;
};

export type TimestampsControllerCreateResponse =
  TimestampsControllerCreateResponses[keyof TimestampsControllerCreateResponses];

export type TimestampsControllerRemoveData = {
  body?: never;
  path: {
    id: string;
  };
  query?: never;
  url: "/timestamps/{id}";
};

export type TimestampsControllerRemoveErrors = {
  /**
   * Timestamp id must be a UUID.
   */
  400: unknown;
  /**
   * Authentication required.
   */
  401: unknown;
  /**
   * Timestamp not found.
   */
  404: unknown;
};

export type TimestampsControllerRemoveResponses = {
  /**
   * Timestamp deleted.
   */
  204: void;
};

export type TimestampsControllerRemoveResponse =
  TimestampsControllerRemoveResponses[keyof TimestampsControllerRemoveResponses];

export type TimestampsControllerFindOneData = {
  body?: never;
  path: {
    id: string;
  };
  query?: never;
  url: "/timestamps/{id}";
};

export type TimestampsControllerFindOneErrors = {
  /**
   * Timestamp id must be a UUID.
   */
  400: unknown;
  /**
   * Authentication required.
   */
  401: unknown;
  /**
   * Timestamp not found.
   */
  404: unknown;
};

export type TimestampsControllerFindOneResponses = {
  200: TimestampModel;
};

export type TimestampsControllerFindOneResponse =
  TimestampsControllerFindOneResponses[keyof TimestampsControllerFindOneResponses];

export type TimestampsControllerUpdateData = {
  body: UpdateTimestampModel;
  path: {
    id: string;
  };
  query?: never;
  url: "/timestamps/{id}";
};

export type TimestampsControllerUpdateErrors = {
  /**
   * Timestamp id must be a UUID.
   */
  400: unknown;
  /**
   * Authentication required.
   */
  401: unknown;
  /**
   * Timestamp not found.
   */
  404: unknown;
};

export type TimestampsControllerUpdateResponses = {
  200: TimestampModel;
};

export type TimestampsControllerUpdateResponse =
  TimestampsControllerUpdateResponses[keyof TimestampsControllerUpdateResponses];